Fix #7: Add workaround for use with previewSite
Jetty as used by unfiltered in sbt-site's previewSite task doesn't
normalize paths before looking up the resulting file. In the case of the
search index the base URL was "./" which was concatenated with the
absolute search index URL "/mkdocs/search_index.json" leading to
".//mkdocs/search_index.json" which resulted in

    GET http://localhost:4000//mkdocs/search_index.json 404 (Not Found)

Workaround this issue by using "." as the base URL and prefixing search
index locations with "/".

The workaround has been tested with serving the theme's site using both
`previewSite` and `python -m SimpleHTTPServer 8000`